refactor: combine util and spec to single file
